Fix: wrong relative links for RSS type and description
The table under `createCollection()`section has invalid relative links for the same pages content. This is due to moving the `reference/` directory from the root of the project to under `docs/` directory.
This is a fix for the RSS relative links by adding `/docs` to the RSS and docs links row in that table.